+++ arelle/ModelDocument.py
@@ -19,7 +19,8 @@ from arelle.PrototypeDtsObject import LinkPrototype, LocPrototype, ArcPrototype,
 from arelle.PluginManager import pluginClassMethods
 from arelle.PythonUtil import OrderedDefaultDict, normalizeSpace
 from arelle.XhtmlValidate import ixMsgCode
-from arelle.XmlValidate import VALID, validate as xmlValidate, lxmlSchemaValidate
+from arelle.XmlValidateConst import VALID
+from arelle.XmlValidate import validate as xmlValidate, lxmlSchemaValidate
 from arelle.ModelTestcaseObject import ModelTestcaseVariation
 
 creationSoftwareNames = None
