+++ arelle/plugin/validate/EFM/MessageNumericId.py
